ExpandButton fixes:   * fix flicker at beginning of animation   * code cleanup
TODO:
Weird flicker after the button has closed again, which is a repaint problem issued by hiding the widget we paint. I'm not sure why this happens. Qt seems to redraw some weird crap for a short moment, instead of the correct rectangle of the parent-widget. Let's hope this is not a Qt/X limitation. If it is, I'm fed up with these shenanigans and will remove the animation completely until we have a proper X-server.

CCMAIL: 75356@bugs.kde.org

svn path=/trunk/kdeextragear-1/amarok/; revision=289302