Made amaroK aRts-independent. Using the --without-arts configure switch it is now possible to build amaroK without aRts support, utilizing only the NMM or GST engine.
This should be good news for all those users who cannot or do not want to install the whole kdemultimedia shebang just to build amaroK. Currently NMM and GStreamer support is not on par yet with our aRts-engine, but we are constantly working on improving this.

amaroK is now becoming a testing platform for future KDE multimedia technologies. Any help with this plan is welcome.

CCMAIL: dkite@shaw.ca

svn path=/trunk/kdeextragear-1/amarok/; revision=294887