Renamed TitleProxy class to "StreamProvider", and added documentation.
This change was done for reflecting the goals of this class, which has shifted it's focus recently. The StreamProvider connects to streaming radio servers and initiates stream transmissions, then processes  the streams and provides amaroK with audio and meta data.

The class has two modes of transferring stream data to the application:

* Signal: Transfer the data directly via Qt SIGNAL. (this mode is preferred).
* Socket: Sets up a proxy server and writes the data to the proxy. This mode is needed for aRts.

CCMAIL: amarok-devel@lists.sourceforge.net

svn path=/trunk/kdeextragear-1/amarok/; revision=332089