Merge in crossfading code for Gst-Engine. This is pretty complicated stuff, so expect some breakage.
Design:

This crossfader implements an "n-track crossfader" design, which means, there's an unlimited number of channels available for mixing. There is one output pipeline, which is never destroyed, and n input pipelines, which are dynamically created and destroyed. Each input has a separate volume element, for defining the percentage of volume which it delivers into the output.  The last stage of the input pipeline is a queue element, which buffers the data in case of pipeline stalls. The bigger the queue is, the less dropouts you get, but you will also see increased latency.

All input pipelines are linked to an adder element in the output pipeline, which combines the signals before handing over to the output sink.

Have fun :)

svn path=/trunk/kdeextragear-1/amarok/; revision=341758