Patch from Sebastian Strand <coffeemonkey@halfnull.net>:
I discovered a problem with the logic in the playlist.cpp setFilter function.
The filter logic only took into account the last of the given terms. The
reason that this often worked is that it also only filtered the visible items
when extending the previous query, meaning that when a query is typed
incrementally the playlist gets filtered correctly. However, manipulate the
query in some other way (like inserting words in the middle) and it fails.

svn path=/trunk/kdeextragear-1/amarok/; revision=384702