Patch from Daniel Winter <wint@gmx.de>:
i made two changes to the coverfetcher.cpp from the cvs version.

The first one uses type=lite instead type=heavy for the webservice query
from amazon.com
That reduces traffic for the cover fetching.

And i changed the xml parsing so that the Name of the Artist is
displayed before the name of the Album in the coverfetching dialog.
That is very usefull because i had often problems to know if the shown
cover is right (when there was two albums with the same name from
different artists).

svn path=/trunk/kdeextragear-1/amarok/; revision=406789