SmartPlaylists are editable!!!
Using XML we store the options the user set in the dialog (together with the SQL query, so that we don't need to parse the xml all the time), and then the user can edit them later.

Dropped support for old format, as it's impossible to parse the SQL back into something editable.

No sqlfortags/sqlforurls anymore, now all smartplaylists behave the same way (using sqlForTags), so we don't need to check this all the time.

BUG: 91036

svn path=/trunk/extragear/multimedia/amarok/; revision=416639