A working, useful configuration dialog for the helix engine.
Allow a user to specify the path where to find the Helix or Realplayer 
installation. Care was taken to allow the user to change the paths, even if
it fails to initialize at first.  Direct the user to the Engine dialog to
fix the paths if the Helix or Realplayer libs cant be found.

The paths are saved in amarokrc, so that they will be remembered for the next
invocation.  No need to create links in the filesystem anymore.  And there was 
great rejoicing in the land...

Look for and load the Helix or Real logo, if found in the Helixplayer or 
Realplayer installation.

If engine successfully loads, the Helix/Realplayer engine is queried for it's
list of plugins, their copyright info, and more info url.  A tab is added to
the config dialog that allows the user to see which plugins are available.
No sophistication here; just because the plugin is shown doesnt mean we can
use it for anything.  For example, what would we do with the Macromedia Flash 
plugin? 



svn path=/trunk/extragear/multimedia/amarok/; revision=433340