Current playlist is now being remembered when amaroK crashes.
The cool thing is that I managed to do this with with virtually no performance penalty. Works like this: We save the current playlist after each modification (Items added, removed..). Then we use the same file for the Undo mechanism, by simply copying it. This way the expensive XML construction only needs to be done once.

Watch out for regressions.

BUG: 98689


svn path=/trunk/extragear/multimedia/amarok/; revision=434213