Implemented my own SW Volume interface.
Why? Because the helix sw volume interface is apply
before the post-mix hook, thus, the scope amplitude 
changes with the volume when SW volume is enabled
(and goes dead when the player is muted)

The volume is now applied after the scope is fed,
and after equalization.  Now the scope is constant
and continues to display when the player is mute.

The old behavior can be reenabled at compile time by 
defining HELIX_SW_VOLUME_INTERFACE in helix-sp.


svn path=/trunk/extragear/multimedia/amarok/; revision=438297