set the OpenAudioDeviceOnPlayback to 0.
Yes, that's right, all this code was required to do that...

This behavior is preferable to having the PCM volume set back to 50% on every
track change, but is not perfect - the PCM volume will be reset now to the
initial value on startup.  At least you can set one value and use SW volume 
control after that.

Still working on having it just not muck around with it at all.
 

svn path=/trunk/extragear/multimedia/amarok/; revision=454849