Fix gst-engine troubles with alsasink (this time for real).
Reading the decodebin example code again, I found that the sink should be in a separate bin,
which then must be added to the thread in the newPad callback. Apparently alsasink does not
like to be fired up without a working source linked to it.

CCMAIL: amarok-devel@lists.sf.net


svn path=/trunk/extragear/multimedia/amarok/; revision=455349