The new Collection Scanner goes live :) The design is as follows:
The scanner itself is run in an external process, unlike before, where it used to be thread. The advantage is that the scanner cannot crash the amaroK main application any more. If it crashes we can simply restart it. 

amaroK communicates with the scanner via the ScanController class, which processes XML entities written to stdout by the scanner process. For XML parsing an event driven SAX2 parser is used, which can process the entities as they arrive, without the need for a DOM document structure.

Currently the new scanner still has some limitations, which will be amended soon:
* Incremental scanning not yet implemented. For this the old scanner is still used.
* No playlist importing.
* No cover image reading.
* No tag recoding. (yikes, let's get rid of that already)

CCMAIL: amarok-devel@lists.sf.net


svn path=/trunk/extragear/multimedia/amarok/; revision=476874