If you are going to implement reference counting, you really should INITIALIZE the count.
It's amazing how much more effective it is.

Also, here we learned that RemoveStreamInfoResponse releases the 
StreamInfoResponse.  When the reference count is initialized correctly
(as is the case in gcc < 4, which is what Untouchable uses), calling 
Release again causes a double delete, which is bad...


I think we got this one now.


svn path=/trunk/extragear/multimedia/amarok/; revision=483305