fix regression caused by moving the scope/equ/vol hook to a FinalHook.
Apparently the packet time in a final hook is a "presentation time", in
other words it tracks how long you've been playing the stream - which is fine
unless you seek forward or back in the stream.  Visualizations would stop
updating following a sync. 

Since we switched to this hook so that visualizations would track better
while playing live (radio) streams (where seeking is not relevant), the 
solution was to feed visualizations from a final hook when playing a live
stream, and from a postmix hook when playing a local one.

Also updated the time tracking in the visualizations to account for both 
players active simultaneously during a crossfade.

Finally, tell the helix core into using the PC Speaker mixer for volume
control when using alsa.  Since we do volume control ourselves, the purpose 
is to fool the core into using a mixer that not many will care about, so that
when it gets reset to its initial value it wont be so annoying. (experimental)


svn path=/trunk/extragear/multimedia/amarok/; revision=498315