Tools menu now contains "Medium Plugin Manager"  :-)  Might need a new name, definitely a new icon.
Anyways, you can now change plugins for media devices.  HAL IDs used, of course, though device names shown in dialog.  
Writes to config file as expected.  Dialog is ugly, as is the original first-time plugin chooser, but comments (or code) on 
making it nicer are welcome.

sebr:  Yes, the code is ugly as crap.  Was going for functionality, as I'm way late in getting this out the door.  I will 
go back at some point and clean it all up.  I promise :-D


svn path=/trunk/extragear/multimedia/amarok/; revision=499136