Create a nice, smooth, reliable crossfade in the helix engine.
As with xine, requires HW mixing or a working dmix.

This is accomplished by creating 1 player in each of 2 child processes and 
controlling them through pipes.  an anonymous mmap is used to for time-critical
current time of each player and for visualization buffers (to reduce cpu).

works well here, but this is a fairly significant rework of the internals of 
the helix engine, so please test and report bugs!



svn path=/trunk/extragear/multimedia/amarok/; revision=511770