* Improved handling of embedded cover art, utilizing the database. Patch       by Shane King <kde@dontletsstart.com>.
Quoting the submitter:
The collection scanner stores information about what art is avaiable in a music file in the database. The art is identified by its MD5 sum, so that duplicates are stored as a single file. This also fixes the problem of art going out of date, since when the art changes, so does the MD5. Due to the information about what art is available being in the database, it can be associated correctly with albums for the other albums sidebar and the album art browser. 

Additionally it allows for multiple pieces of art per file, and stores the embedded description of each, although nothing is done with this information at the moment (much like multiple image files always selects only one to display).

BUG: 124563


svn path=/trunk/extragear/multimedia/amarok/; revision=524522