Many changes and improvements.  Manual device management is pretty much at completion, except for per-device mount/unmount/transcode settings.
Current known bugs:

1.  Sometimes devices that are configured will be detected as new.  This seems to be a bug in KConfig, and
has not yet been reported by anyone else, so will not be looked into unless it is.

2.  When clicking Autodetect, it detects both manually-configured devices and HAL/DBUS autodetected devices.  This should
be fixed (maybe by checking if a device is in the delete map when detecting?)

Also, a new dcop call that displays the current device list as known to the Device Manager (both autodetected and manual
devices).


svn path=/trunk/extragear/multimedia/amarok/; revision=527153