New option:  [General]->AppendAsDefault
If true, sending new media to amaroK (i.e. via ampache) will do an append to the current playlist.  Useful for i.e. Firefox, where you can 
tell it to send M3U files to amaroK, but are unable to specify the -a flag so your current list is not overwritten.  Will GUI-fy it into 
Options somewhere once string freeze is over, for 1.4.1.

svn path=/trunk/extragear/multimedia/amarok/; revision=539006