Change the way fields beginning with "The" are treated in the Collection Browser as discussed on the mailing list.
New behaviour:
"The Who" is displayed as "The Who" but is sorted as if it were "Who, The". It is also
counted as being under the heading "W", rather than "T". This happens only for artists -
other fields are completely untouched.

There are still places that do mangle the string (such as the Cover Manager), but they
will take a bit more work to change.

Personally, it's looking a bit weird to me at the moment, but I'm sure I'll get used to
it.

CCMAIL: amarok@kde.org

svn path=/trunk/extragear/multimedia/amarok/; revision=544368