as discussed on the ML...
move the 1.3 series StreamProvider class into the gst10 engine.
port the streamsrc from gst 0.8 to gst 0.10

limits us to shoutcast live streams, but at least we have metadata for them.

To revert back to gst builtin streams, comment out -DGST_KIOSTREAMS in 
Makefile.am.  We will *not* make this a configure script option...

oh, and by the way, it works :)


svn path=/trunk/extragear/multimedia/amarok/; revision=550049