Speed up ATF by using the entries in the database whenever possible instead of always re-reading the metadata.
Note: I have yet to verify that this code works as intended.  I'll test ASAP, but I believe it should be correct.  Hopefully I've 
introduced no new bugs...  :-)

Martin -- can you give it a go with your 6300-song playlist and let me know how much of a speedup (if any) is gained with this code?  
Thanks!

svn path=/trunk/extragear/multimedia/amarok/; revision=565996