Speed up playlist XML writing approximately 10 times. This patch considerably improves Amarok's overall performance with big playlists and also cuts down on shutdown time.
Details:
Problem was that Playlist::saveXML() would invoke methods like MetaBundle::score(), which accessed the database in many cases. This database access was extremely slow, so we circumvent it and use cached values instead.

Another optimization is that we now write the XML data to a string before writing to the target file.

CCMAIL: amarok@kde.org

svn path=/trunk/extragear/multimedia/amarok/; revision=580555