Relocate VACUUM to after a collection scan has been run and it has produced some changes (i.e. for an incremental scan, the mtime on a directory must have changed). Also, run VACUUM on PostgreSQL.
For PostgreSQL, VACUUM does not cause any locks, but it does for SQLite. This hopefully
should not cause a problem, as long as VACUUM takes less than 12 seconds, as that is the
amount of time we retry for before aborting an SQLite query.

Although, maybe there needs to be a better way to do this.

svn path=/trunk/extragear/multimedia/amarok/; revision=586757