Make AFT emit fileAdded and fileDeleted in manageable ways for both incremental and full rescans.  Make the Playlist use these to do really cool and magic things with files when they're added to and deleted from the Collection and entries already exist in the Playlist, without too much overhead.
To do this I made a change in how PlaylistItems are enabled and disabled, splitting reasons for it.  This makes it far more robust but has the side effect 
that I may have screwed up Dynamic Playlists a bit.  I'll have to do more testing, but as we have no release imminent we have time to catch it.

I'll send an email to the devel list regarding the changes I made, why it was done, and what it means.  In a couple of hours.  Meanwhile, everything other 
than Dynamic Playlists has been tested a bit and seems to be working well.

svn path=/trunk/extragear/multimedia/amarok/; revision=589915