Fix loading directories in a browser from the tag editor if the path contains a parenthesis.
Thanks to Richlv for reporting it, and to hurra for a perfectly good patch fixing the
issue that I didn't use in the end (because Firefox confused me).

It's still broken though if you use Firefox and already have a browser window open. I
blame Firefox's input parsing, and think it doesn't like C-style escaped parentheses (with
an open browser window)

BUG:132961

svn path=/trunk/extragear/multimedia/amarok/; revision=596032