AFT support for lyrics!  Yay!
Also, a bit of AFT refactoring, and an update of the previous change I made to metabundle, so AFT is still not run on those, but the UID in the playlistitem
reflects the correct value for any updated metadata.

Max/Alexandre: Your commit and mine conflicted (I haven't updated in a few days until now) -- I figured out an easy way around Rich's crash too, but I decided to 
go with my way, since it will get the UID in a way that will still be safe and current (at least, moreso) if a scan is ongoing.  Thanks for looking at it though!

svn path=/trunk/extragear/multimedia/amarok/; revision=603770