BUG: 138567
This commit shifts currentFolder declaration inside the conditional
scope that it is actually used in and removes an unneeded call to set
currentFolder to the current folder.

This commit also removes an unused setting of an iterator and moves
the declaration of the iterator to be inside the for loop scope.

svn path=/trunk/extragear/multimedia/amarok/; revision=613511