BUG: 138556
This commit shores up the few places I found where a function is
returning a value and the code is not checking it. Mostly it is
just reporting the value if things go wrong. But in a few places
the caller will take an action based on the return value.

svn path=/trunk/extragear/multimedia/amarok/; revision=613522