BUG: 138564
The MetaBundle class is a touch large. Passing it around by value is a
bad idea if we don't have to. This commit changes the dcop use of
MetaBundle to be pass by reference instead. The performance boost
is marginal at best. But the code worked out to be pretty easy to
do since the code already did nothing to the MetaBundle object in use.

svn path=/trunk/extragear/multimedia/amarok/; revision=613527