Bunch of things in this commit:
1) Support for colored stars *everywhere*.  Works in Playlist, OSD, and Context Browser.  (Can't pick your own colors yet, but that's coming...for now, 
committing to get things tested, since this is all the under-the-hood support).

2) Bugfix: OSD wouldn't have rating updated if rating of current song changed in playlist while OSD displayed.

3) Bugfix: Context browser wouldn't have rating of suggested songs or songs by same artist be updated if changed in Playlist.  Now it is.  If this causes 
(unacceptable) slowness (which I don't think it does), it can be reverted, but I think it's more intuitive and it looks more professional and polished.

4) Change: Rating is represented in some places as 1-10 (10 levels) but in others as 0-5, but in places where 0-5, couldn't select 0.5 (so only 9 levels).  
So there was no way to ever assign a rating of 1 and a weird disconnect between the two ways of representing it.  Now assigning a single half-star to a 
track is valid (1/10 or 0.5/5).  Can still clear ranking by editing tag and manually entering  0, but I think most people that assign a rating will want to 
keep a rating on the track, and it always seemed very unintuitive that there was a "missing" rating value.


svn path=/branches/stable/extragear/multimedia/amarok/; revision=633748