Merge the uirefactor branch back to trunk.  This branch explores the concept of a three panel layout, a panel dedicated to finding music--a panel dedicated to managing music--and a panel dedicated to viewing information about ones music.  This layout is designed on a mockup done by Max Howell.
Enhancements: Amarok is now a "proper" main window.  This should allow for easier ui redesign in the future if it should become necessary.  The context browser has become a central widget that is much more central to the playback experience.  This enhances the important of the context browser, or context pane as it is now called.  

REGRESSIONS: * Browsers cannot be hidden properly.  Because we now have to add widgets to a QDockWidget before adding them to the ui, when the widget is hidden the dock widget remains--this needs attention.
* The statusbar looks extremely crowded.  The information needs to be broken up, with the song progress bar making it into the top toolbar and the progress dialog being moved to the browsers pane. 
* The playlist also looks very crowded.  This should be fixed as the playlist gets refactored to fit the new UI.

This commit also includes a few other nonrelated changes--Using Qt::escape() instead of our own escapeHTML; cleaning up some header files, and experimental support for video playback (disabled).

svn path=/trunk/extragear/multimedia/amarok/; revision=638175