Debug output is now a runtime option (--debug), as opposed to compile time.
The way I'm suppressing the debug output from QDebug is a bit of a hack, but it works and I don't know another way. Another problem is
that adding or removing the debug option doesn't immediately show effects on the first startup, because the option is only read after Amarok is mostly initialized. But oh well, too tired to think about that now.


svn path=/trunk/extragear/multimedia/amarok/; revision=697123