Delete columnlist, which dealt directly with the old playlist.  Now we can see what doesn't work a lot easier.
The completes deleting the old playlist and supporting files.  Anything that was explicitly disabled due to incompatibility with the new playlist at this point has been noted with a // PORT 2.0 or a //TODO: PORT.

I have left the old dynamicbar/dynamic title code in mainwindow, and dynamicmode.{h,cpp} around as theres still a great deal of logic in there that will be helpful in reproducing dynamicmode for the new playlist.

svn path=/trunk/extragear/multimedia/amarok/; revision=708467