Fix the playlist constantly redrawing itself using 100% cpu time. required a near complete rewrite of the paint() method.
It turns out that there are quite a few "gotcha's" when inside the paint method of a QGraphicsItem. Any call to show() of a child item or setBrush() WILL cause paint to get called again. This is ok if one is very carefull that 
the same call will not occur from within the next call to paint(), as otherwise we have an infinte, if non blocking, loop eating up resources as fast as it can.

I managed to somehow kill the text of the playlist items, how is beyond me at the moment, but I will work on bringing it back!


svn path=/trunk/extragear/multimedia/amarok/; revision=710712