Implement sorting in the collectionbrowser by Track Number.
We reimplement lessThan in the proxyModel to do so.  Note that this sort order does not propogate down to the source model, which means that adding items to the playlist from the collection does not insert them in the sorted order.  I'm not quite sure the proper way to fix this at present, will investigate.

svn path=/trunk/extragear/multimedia/amarok/; revision=726580