No longer fetch images by default in Meta::Album::image()
Instead we do it in the playlist and collection browser.  This allows the cover manager to fetch on demand rather than on loading it.  Remove a bit of debug from playlistgraphicsitem as it was being quite spammy.
Maxx: I think your earlier commit caused a loop in PlaylistGraphicsItem::resize(), as its being called about a million times a second.
CCMAIL: maximilian.kossick@googlemail.com

svn path=/trunk/extragear/multimedia/amarok/; revision=743847