First of all, yes, unfortunately this commit destroys quite a bit of history in the scriptable service directory, but as nothing in there has really been usefull up to this point ( I wrote it all, so I think I am a good authority on that ), I dont think this is a major loss...
So, what does this huge commit actually do?

1. Add a redesigned scriptable service with a sane DBus API consisting of only 3 functions. Make amount of levels 
configurable and make everything completely dynamic
2. Make the service scripts use the same callback method as regular scripts
3. Integrate scriptable services with the service manager, making it possible to run and stop service scripts from 
there
4. Add a new service category to the scriptmanager
5. Add ( recreate ) a simple cool streams service script. This one works really nicely
6. Add a deme file browser script. This one is mainly for testing the scriptable services and currently contains some 
hardcoded paths, so it should not be included in a release ( it is also basically useless, but I had to have something 
to test agains when I had no internet )
7. Fix a ton of bugs and crashes in the new scriptable service.

All in all, not bad for a few days in a small wodden cabin with a girlfriend who is busy studying for an examn... :-)  

And to all the people I have been telling that there would be no scriptable service in 2.0, sorry, but I changed my 
mind!


svn path=/trunk/extragear/multimedia/amarok/; revision=788838