Some one on enginecontroller:
Use a more phononesque approach, rather than having the playlist control what the engine is doing, have the engine grab the next tracks from the playlist and queue them for phonon to play.  This involved changing the track navigators.In addition, only tell phonon to seek if we give it a specific point in the file to seek to.  phonon/xine is really bad at seeking.

svn path=/trunk/extragear/multimedia/amarok/; revision=789337