remove hydrogen's stuff
Removed Hydrogen's hack which was supposed to make the Progress-Slider smooth; this hack could never have worked correctly.
Instead, I made it so that the engine sends out timer ticks with a shorter interval, which gives us more precision, and it also eliminates the weird bugs.

Additionally I made the code simpler and cleaner.

svn path=/trunk/extragear/multimedia/amarok/; revision=810808