Huzzah, Amarok 2 gets video support! (for music videos only. Amarok isn't meant to replace your video player!).
Added a video applet for the context view.
While it's not currently possible to show video in a QGraphicsProxyWidget, I made this work by using a normal QWidget and syncing the position with the plasmoid. Turns out this works pretty well.

So we should soon be able to watch youtube videos from SeeqPod, e.g. :)

Oh yeah, work-in-progress of course.

svn path=/trunk/extragear/multimedia/amarok/; revision=812620