It is time to introduce SqlPlaylist and friends to the world!
The SqlPlaylist is currently a very simple playlist that stores itself in the Amarok database. It can currently store and load tracks from the database but not much else. In support of the 
SqlPlaylist, a basic playlist browser has been implemented. This browser supports an unlimited number of playlist grouped in an arbitrary hirachy of grups ( or virtual folders ). Support for the 
groups has not been exported to the gui layer yet though.

The old save_playlist action has been renamed to export_playlist, and save_playlist is not used for storing a playlist in the database. An action for this has been added below the playlist browser, 
and using this actually works :-) What currently does not work, however, is readding a stores SqlPlaylist to the playlist as that crashes


svn path=/trunk/extragear/multimedia/amarok/; revision=818496