Fix slight flicker when resizing left sidebar.
Create browser widgets in sidebar with a parent widget
which belongs to the top-level main window.  Otherwise
various Qt/KDE calls (eg. QWidget::show(), KDirOperator::setDirLister())
made while the widget does not have a parent will result in the widget 
being 'promoted' to a native X11 window which will flicker slightly when
resized.  The widgets which an X11 window is being created for can be 
determined by setting a breakpoint on qtXCreateWindow and qtXCreateSimpleWindow
on startup.

Parent the Playlist::GraphicsView widget to the mainWindow() upon creation.


svn path=/trunk/extragear/multimedia/amarok/; revision=830509