Massive patch changes the name of Meta::Track::url to Meta::Track::uidUrl.  From Max:
"Meta::Track::url() isn't anmed correctly anyway, but I'm too lazy to refactor it. Look at the documentation for that method, it *is* a uid, it's just not named correctly. the return value of that 
method may only be used within amarok. the sqlcollection plugin wuld have to interpret amarok-sqltrack urls, so it would be free to use some magic to figure out if the url contains a uid or a path."

I decided to fix his laziness :-)

This helps reinforce that it's a track specific unique ID method; in SqlCollection this uidUrl will eventually contain the AFT ID, i.e. amarok-sqltrack:/273ab6s5c5d6d or some such thing.  In fact, 
a portion of this (if not the mechanism that does the actual tracking) will be implemented soon, because the UID part of AFT is needed by xevix to do proper collection syncing with the iPod.

CCMAIL:amarok-devel@kde.org


svn path=/trunk/extragear/multimedia/amarok/; revision=840672