I'm glad to present you the toolbox menu (it appears after clicking on the add icon). It should replace the applet browser which I still haven't removed entirely (right click->Add applet...) so for now they should coexist for a little while. There are still some problems (placement mostly) that should be fixed by a lot of refactoring in the ContextView class that I hope will fix also another issues.
For now you can test the menu a little bit, it is still slow and buggy but I hope I can fix it. Also it's not possible to remove applets from it but I'm unsure if that is what we want/need. If you click in an entry it adds an applet or changes to the containment where the applet is if it was already added.

Some slightly changes in the toolbox look and also some refactoring.


svn path=/trunk/extragear/multimedia/amarok/; revision=842813