AFT WORKS!  For parts of it at least.  Move your files around and watch as the playlist automagically keeps working  :-)
Next up is getting lyrics and statistics back up, although that should be relatively trivial.

One issue currently: the behavior for the OSD when notifyObservers is called is to pop up.  This should probably be changed, as it's probably not what we actually want -- if 
the currently playing track is the one that was changed, only in that case should it pop up.

svn path=/trunk/extragear/multimedia/amarok/; revision=843349