playPause() isn't the only way Amarok can be paused or start playing.
Instead, listen to Phonon state changes.  This means that the signal is triggered every time the track changes as well, but I can't see a way around that short of keeping track of the current state and comparing it (and doing the same for "stopped").

Note that we treat buffering as playing.


svn path=/trunk/extragear/multimedia/amarok/; revision=853403