Enable MySQL Embedded support. So now Amarok 2 requires the libmysqld.a library. On OpenSUSE, this is installed with libmysqlclient-devel, YMMV.
We're pretty close to the beta2 tag, so if something is commonly not working
ping me and I'll revert. It'd be really nice to have this tested in the beta2
though.

The reason for the 'FindMySQLAmarok' file is that the default FindMySQL couldn't
find mysql embedded on my system. I figure there might be some more corrections
needed, then we can submit our changes upstream in ~1 month.

CCMAIL: amarok@kde.org
CCMAIL: amarok-devel@kde.org

svn path=/trunk/extragear/multimedia/amarok/; revision=863921