Major overhaul of the playlist subsystem
== Changes for end users ==
In the playlist view, mouse and keyboard event handling, and the selection
system have both been significantly improved.  The undo/redo commands now work
properly in all cases.  Album grouping works consistently.  Some of the eye
candy (like animations) has been temporarily removed, but will hopefully return
in the future.

== Changes in general Amarok code ==
The PlaylistModel has been split into four different classes:
1) PlaylistActions
2) PlaylistController
3) PlaylistModel
4) GroupingProxy

The play, nextTrack, and previousTrack functions are now part of
PlaylistActions.  Functions that modify the contents of the playlist (insert,
remove, and move tracks; undo and redo) are now part of PlaylistController.  If
you want to find out about the contents of the playlist, you still use
PlaylistModel, though you cannot modify it without using the Controller.  I
encourage you to take a look at the headers for each of these classes to find
out what functions are available.

The album grouping functionality that used to be in the PlaylistModel has been
moved into GroupingProxy.  Both GraphicsView and PrettyListView (described in
the next section) use GroupingProxy.  It really isn't useful for anything other
than playlist views.

The signals and slots for the PlaylistModel have changed substantially;
PlaylistModel now behaves much more like a QAbstractItem model when it is
modified.

A position-independent accessing system has been added to the model for
situations where you care what is in the playlist, but not what order the
playlist is in (eg, Random Track mode).

The Model::trackNumberLessThan() function has moved to Meta::Track::lessThan()

== Changes in the playlist subsystem ==
The src/playlist/ directory has been rearranged.

Along with the PlaylistModel changes, there is a new View for the playlist,
called PrettyListView.  It subclasses QListView and is designed to be a
replacement for GraphicsView.  The code for GraphicsView still exists in the
source tree and has been mostly ported to the new PlaylistModel system, but
PrettyListView solves a lot of the selection model and event handling problems
of the GraphicsView.

CCMAIL: amarok-devel@kde.org
CCBUG: 167861
CCBUG: 171048
CCBUG: 171622

svn path=/trunk/extragear/multimedia/amarok/; revision=868947