Fixed various "crash on remove tracks" bugs in PlaylistModel
Damn, that one took a while to figure out what was going on.  The
problem was that my clever insertRows/removeRows strategy for the
PlaylistModel was screwing up the internal state of the GroupingProxy.
The ListView holds PersistentModelIndexes to various parts of the
GroupingProxy, and when the GroupingProxy and the PlaylistModel got
unsychronized, Amarok would crash pretty soon thereafter.

So I reworked the insertRows and removeRows approach, and cleaned up a
few problems in other places.

CCMAIL: amarok-devel@kde.org

svn path=/trunk/extragear/multimedia/amarok/; revision=869326