Unhide function parameters.
It is legal in C/C++ to declare variables within functions
the same name as parameters of the function. But it is dangerous
to do so since it hides that parameter within the scope the
declaration was made. There were two instances of this in Amarok.
This commit renames those declarations to reveal those parameters.

svn path=/trunk/extragear/multimedia/amarok/; revision=870980