Beginnings of a first run tutorial.  Idea below.  Yes, this will massively break string freeze.  Yes, it is worth it.
What the code currently does is, if you have FirstRunTutorial=true in the General section of your amarokrc, it will upon startup fade a semi-transparent overlay on top of the main window, then after a few seconds kill itself.

Here's my vision (read the comments in FirstRunTutorial.cpp too): Bottom right has next and prev buttons (like QGraphicsWidgets with QPushButtons); top right has close button.  Next and prev flip through different "pages" of the tutorial.  (Code has some ideas as to how to make this as painless and code reuse-ful as possible).  Arrows pointing to relevant parts of the window along with descriptive text fade in and out as pages transition.  Perhaps some buttons to launch a 1.4 database importer and/or collection setup as well.

For each page, have a description of a feature of Amarok and an arrow pointing to the relevant place.  For instance, an arrow pointing to the left sidebar at the level of the Internet tab would then describe what's inside of that tab.  There's also no reason that pages couldn't activate various parts of Amarok as well, such as describing a tab, then on the next page activating it and describing subcomponents.

Leo and Seb and I think a few more are totally into the idea and think it's way worth breaking string freeze.  I'm committing it now because my time this upcoming week to hack doesn't look great, so I'd love help, even though I'd have loved to have it a bit further so that it'd be more obvious what I was thinking (hopefully the text above will be enough instead  :-)  )

CCMAIL:amarok-devel@kde.org


svn path=/trunk/extragear/multimedia/amarok/; revision=871164