Replace macro emitProperResult with template function emitProperResult.
Since we are using C++, there is no reason to have a macro designed to
deal with different types. Instead we should attempt to use templates
for the type generality and make the macro a function. This leads us to
the macro emitProperResult. It is used in 6 different source code files
as a way to determine the right thing to return even when the list list
that the thing is gotten from is different. Great place for a template
instead.

This commit replaces each instance of the macro with a member function
for the class that uses where the member function is a templated
function. It also attempts to reconcile the different classes so that
the template function looks very much the same. This is to make it
easier to identify these functions so that some industrious young lad or
lass can go in and squash the 6 different member functions into about 2
(MemoryQueryMaker is different enough that it will require its own
function).

BUG: 172755

svn path=/trunk/extragear/multimedia/amarok/; revision=872244